+++ packages/core/src/render3/component_ref.ts
@@ -14,9 +14,6 @@ import {
   NotificationSource,
 } from '../change_detection/scheduling/zoneless_scheduling';
 import {Injector} from '../di/injector';
-import {convertToBitFlags} from '../di/injector_compatibility';
-import {InjectFlags, InjectOptions} from '../di/interface/injector';
-import {ProviderToken} from '../di/provider_token';
 import {EnvironmentInjector} from '../di/r3_injector';
 import {RuntimeError, RuntimeErrorCode} from '../errors';
 import {DehydratedView} from '../hydration/interfaces';
@@ -32,7 +29,6 @@ import {NgModuleRef} from '../linker/ng_module_factory';
 import {Renderer2, RendererFactory2} from '../render/api';
 import {Sanitizer} from '../sanitization/sanitizer';
 import {assertDefined, assertGreaterThan, assertIndexInRange} from '../util/assert';
-import {NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR} from '../view/provider_flags';
 
 import {AfterRenderEventManager} from './after_render_hooks';
 import {assertComponentType, assertNoDuplicateDirectives} from './assert';
@@ -91,6 +87,7 @@ import {mergeHostAttrs, setUpAttributes} from './util/attrs_utils';
 import {debugStringifyTypeForError, stringifyForError} from './util/stringify_utils';
 import {getComponentLViewByIndex, getNativeByTNode, getTNode} from './util/view_utils';
 import {ViewRef} from './view_ref';
+import {ChainedInjector} from './chained_injector';
 
 export class ComponentFactoryResolver extends AbstractComponentFactoryResolver {
   /**
@@ -157,40 +154,6 @@ function toRefArray<
 function getNamespace(elementName: string): string | null {
   const name = elementName.toLowerCase();
   return name === 'svg' ? SVG_NAMESPACE : name === 'math' ? MATH_ML_NAMESPACE : null;
-}
-
-/**
- * Injector that looks up a value using a specific injector, before falling back to the module
- * injector. Used primarily when creating components or embedded views dynamically.
- */
-export class ChainedInjector implements Injector {
-  constructor(
-    public injector: Injector,
-    public parentInjector: Injector,
-  ) {}
-
-  get<T>(token: ProviderToken<T>, notFoundValue?: T, flags?: InjectFlags | InjectOptions): T {
-    flags = convertToBitFlags(flags);
-    const value = this.injector.get<T | typeof NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR>(
-      token,
-      NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR,
-      flags,
-    );
-
-    if (
-      value !== NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR ||
-      notFoundValue === (NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR as unknown as T)
-    ) {
-      // Return the value from the root element injector when
-      // - it provides it
-      //   (value !== NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR)
-      // - the module injector should not be checked
-      //   (notFoundValue === NOT_FOUND_CHECK_ONLY_ELEMENT_INJECTOR)
-      return value as T;
-    }
-
-    return this.parentInjector.get(token, notFoundValue, flags);
-  }
 }
 
 /**
