+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -1137,6 +1137,9 @@
     "name": "init_callback_scheduler"
   },
   {
+    "name": "init_chained_injector"
+  },
+  {
     "name": "init_change_detection"
   },
   {
